handle_bot_message/requirements.txt
boto3==1.26.142
pytz==2023.3
botocore~=1.29
gspread==5.9.0
python-telegram-bot~=20.3
